pipeline/video_studio/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
pydantic==1.10.22
requests==2.32.4
httpx==0.28.1
Pillow==11.3.0
matplotlib==3.10.5
trafilatura==2.0.0
beautifulsoup4==4.13.5
readability-lxml==0.8.4.1
postgrest==1.1.1
groq==0.31.0
piper-tts==1.3.0
imageio-ffmpeg==0.6.0
